# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EventType = typeof EventType[keyof typeof EventType];


export const EventType = {
  visit: 'visit',
  symptom: 'symptom',
  medication: 'medication',
  test: 'test',
  diagnosis: 'diagnosis',
  note: 'note',
} as const;

export interface Profile {
  id: number;
  name: string;
  relationship: string;
  initials: string;
  color: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  bloodGroup?: string | null;
  eventCount: number;
  /** @nullable */
  lastUpdated?: string | null;
}

export interface ProfileInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  relationship: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  bloodGroup?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface ProfileUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  relationship?: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  bloodGroup?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface HealthEvent {
  id: number;
  profileId: number;
  type: EventType;
  title: string;
  date: string;
  description: string;
  /** @nullable */
  provider: string | null;
  /** @nullable */
  location: string | null;
  medications: string[];
  tags: string[];
  /** @nullable */
  followUp?: string | null;
}

export interface EventInput {
  type: EventType;
  /** @minLength 1 */
  title: string;
  date: string;
  description: string;
  /** @nullable */
  provider?: string | null;
  /** @nullable */
  location?: string | null;
  medications?: string[];
  tags?: string[];
  /** @nullable */
  followUp?: string | null;
}

export interface EventUpdate {
  type?: EventType;
  /** @minLength 1 */
  title?: string;
  date?: string;
  description?: string;
  /** @nullable */
  provider?: string | null;
  /** @nullable */
  location?: string | null;
  medications?: string[];
  tags?: string[];
  /** @nullable */
  followUp?: string | null;
}

export type ReminderStatus = typeof ReminderStatus[keyof typeof ReminderStatus];


export const ReminderStatus = {
  upcoming: 'upcoming',
  due: 'due',
  completed: 'completed',
  rescheduled: 'rescheduled',
  missed: 'missed',
} as const;

export interface Reminder {
  id: number;
  profileId: number;
  title: string;
  date: string;
  detail: string;
  status: ReminderStatus;
  completed: boolean;
}

export interface ReminderInput {
  profileId: number;
  /** @minLength 1 */
  title: string;
  date: string;
  detail: string;
}

export interface ReminderUpdate {
  date?: string;
  detail?: string;
  status: ReminderStatus;
}

export interface Dashboard {
  profiles: Profile[];
  eventsThisMonth: number;
  upcomingReminders: number;
  recentEvents: HealthEvent[];
  activeProfileId: number;
}

export interface HealthSummary {
  profile: Profile;
  generatedAt: string;
  overview: string;
  keyConditions: string[];
  currentMedications: string[];
  recentTests: string[];
  recentVisits: string[];
}