# lib/api-zod/src/generated/types/reminderStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ReminderStatus = typeof ReminderStatus[keyof typeof ReminderStatus];


export const ReminderStatus = {
  upcoming: 'upcoming',
  due: 'due',
  completed: 'completed',
  rescheduled: 'rescheduled',
  missed: 'missed',
} as const;