# lib/api-zod/src/generated/types/reminderUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReminderStatus } from './reminderStatus';

export interface ReminderUpdate {
  date?: Date;
  detail?: string;
  status: ReminderStatus;
}